pyproject.toml
[project]
name = "good-birds"
version = "0.1.0"
description = "CLI tool to sort and rate bird photography CR2 bursts."
readme = "README.md"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "rawpy>=0.19.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "rich>=13.0.0",
    "click>=8.0.0"
]

[project.scripts]
good-birds = "good_birds.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["good_birds"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
markers = [
    "integration: marks tests as requiring external dependencies like exiftool (deselect with '-m \"not integration\"')"
]
